restructure tests and add tests for explicit container resolvers
