+++ test/functional/tools/explicit_container.xml
@@ -1,0 +1,25 @@
+<tool id="explicit_container" name="explicit_container" version="0.1.0">
+    <requirements>
+        <container type="docker">quay.io/biocontainers/bwa:0.7.17--h7132678_9</container>
+    </requirements>
+    <stdio>
+        <exit_code range="2:" />
+    </stdio>
+    <command><![CDATA[
+bwa > '$output_1' 2>&1 &&
+
+touch "\${TMP:-/tmp}/job_tmp" &&
+touch "\${TEMP:-/tmp}/job_temp" &&
+touch "\${TMPDIR:-/tmp}/job_tmpdir"
+    ]]></command>
+    <inputs>
+    </inputs>
+    <outputs>
+        <data name="output_1" />
+    </outputs>
+    <tests>
+    </tests>
+    <help><![CDATA[
+        TODO: Fill in help.
+    ]]></help>
+</tool>
