+++ test/functional/tools/sample_tool_conf.xml
@@ -264,7 +264,8 @@
   <tool file="mulled_example_invalid_case.xml" />
   <tool file="mulled_example_broken_no_requirements.xml" />
   <tool file="mulled_example_broken_no_requirements_fallback.xml" />
-
+  <tool file="explicit_container.xml" />
+  <tool file="explicit_singularity_container.xml"/>
   <tool file="simple_constructs.yml" />
   <tool file="from_work_dir_glob.xml" />
