+++ test/integration/test_containerized_jobs.py
@@ -1,76 +1,25 @@
 """Integration tests for running tools in Docker containers."""
 
 import json
-import logging
 import os
-import re
 import unittest
-from tempfile import mkdtemp
 from typing import (
     Any,
     Dict,
-    List,
-    Optional,
-    TYPE_CHECKING,
-)
-from typing_extensions import (
-    Literal,
-    Protocol,
 )
 
-from galaxy.tool_util.deps.container_resolvers.mulled import list_docker_cached_mulled_images
-from galaxy.util.commands import (
-    shell,
-    which,
-)
+from galaxy.util.commands import which
 from galaxy_test.base.populators import DatasetPopulator
 from galaxy_test.driver.integration_util import IntegrationTestCase
 from .test_job_environments import BaseJobEnvironmentIntegrationTestCase
-
-if TYPE_CHECKING:
-    from requests import Response
 
 SCRIPT_DIRECTORY = os.path.abspath(os.path.dirname(__file__))
 
 # local_docker
 # local_docker_inline_container_resolvers (using only fallback resolver)_
 DOCKERIZED_JOB_CONFIG_FILE = os.path.join(SCRIPT_DIRECTORY, "dockerized_job_conf.yml")
-DOCKERIZED_JOB_CONFIG = {
-    'runners': {'local': {'load': 'galaxy.jobs.runners.local:LocalJobRunner', 'workers': 1}},
-    'execution': {
-        'default': 'local_docker',
-        'environments': {
-            'local_docker': {'runner': 'local', 'docker_enabled': True},
-        }
-    },
-    'tools': [
-        {'id': 'upload1', 'environment': 'local_upload'},
-    ]
-}
-
-SINGULARITY_JOB_CONFIG = {
-    'runners': {'local': {'load': 'galaxy.jobs.runners.local:LocalJobRunner', 'workers': 1}},
-    'execution': {
-        'default': 'local_singularity', 
-        'environments': {
-            'local_singularity': {'runner': 'local', 'singularity_enabled': True, 'singularity_run_extra_arguments': '--no-mount tmp'},
-            'local_upload': {'runner': 'local'}
-        }
-    },
-    'tools': [{'id': 'upload1', 'environment': 'local_upload'}]}
-
 # define an environment (local_singularity) for local execution with singularity enabled
 SINGULARITY_JOB_CONFIG_FILE = os.path.join(SCRIPT_DIRECTORY, "singularity_job_conf.yml")
-
-JOB_CONFIG_FOR_CONTAINER_TYPE = {
-    "docker": DOCKERIZED_JOB_CONFIG,
-    "singularity": SINGULARITY_JOB_CONFIG,
-}
-
-JOB_CONFIG_FILE_FOR_CONTAINER_TYPE = {
-    "docker": DOCKERIZED_JOB_CONFIG_FILE,
-    "singularity": SINGULARITY_JOB_CONFIG_FILE,
-}
 
 EXTENDED_TIMEOUT = 120
 
@@ -108,106 +57,6 @@ class MulledJobTestCases:
         self.dataset_populator.wait_for_history(history_id, assert_ok=True, timeout=EXTENDED_TIMEOUT)
         output = self.dataset_populator.get_history_dataset_content(history_id)
         assert "0.7.15-r1140" in output
-
-
-class ContainerizedIntegrationTestCase(IntegrationTestCase):
-    """
-    class for containerized (docker) tests
-    provides methods for clearing and checking the container cache
-    cache is cleared before each test
-    """
-
-    container_type: str = "docker"
-    dataset_populator: DatasetPopulator
-    framework_tool_and_types = True
-    jobs_directory: str
-    allow_conda_fallback: bool = False
-
-    @classmethod
-    def setUpClass(cls) -> None:
-        skip_if_container_type_unavailable(cls)
-        super().setUpClass()
-
-    def setUp(self) -> None:
-        super().setUp()
-        self.dataset_populator = DatasetPopulator(self.galaxy_interactor)
-        self._clear_container_cache()
-
-    @classmethod
-    def handle_galaxy_config_kwds(cls, config) -> None:
-        super().handle_galaxy_config_kwds(config)
-        cls.jobs_directory = cls._test_driver.mkdtemp()
-        config["jobs_directory"] = cls.jobs_directory
-        config["job_config"] = JOB_CONFIG_FOR_CONTAINER_TYPE[cls.container_type]
-        # for almost all containerized tests we want to disable
-        # default (conda) resolution
-        # if conda fallback is allowed the conda env is auto installed on tool
-        # execution
-        if not cls.allow_conda_fallback:
-            disable_dependency_resolution(config)
-        else:
-            cls.conda_tmp_prefix = mkdtemp()
-            cls._test_driver.temp_directories.append(cls.conda_tmp_prefix)
-            config["use_cached_dependency_manager"] = True
-            config["conda_auto_init"] = True
-            config["conda_auto_install"] = True
-            config["conda_prefix"] = os.path.join(cls.conda_tmp_prefix, "conda")
-
-    def _clear_container_cache(self):
-        """
-        clear the cached images
-
-        should be overwritten in test classes for singularity
-        """
-        logging.debug("TestMulledContainerResolver.clear_cache")
-        cmd = ["docker", "system", "prune", "--all", "--force", "--volumes"]
-        shell(cmd)
-
-    def _assert_container_in_cache(
-        self, cached: bool, container_name: str, namespace: Optional[str] = None, hash_func: Literal["v1", "v2"] = "v2"
-    ) -> None:
-        """
-        function to check if the container is cached
-
-        should be overwritten in test classes for singularity
-
-        - determines list of cached images using `docker images`
-        - and checks the given name is in the image identifiers of the cached images
-
-        The boolean `cached` sets the assumption in the caching state.
-        `namespace` and `hash_func` are used to filter cached images.
-        """
-        cache_list = list_docker_cached_mulled_images(namespace, hash_func)
-        imageid_list = [_.image_identifier for _ in cache_list]
-        assert cached == (container_name in imageid_list)
-
-
-class SingularityIntegrationTestCase(ContainerizedIntegrationTestCase):
-    """
-    analogous to ContainerizedIntegrationTestCase, just for singularity
-    provides adapted methods for clearing and checking the container cache
-    """
-    container_type: str = "singularity"
-
-    def _clear_container_cache(self):
-        """
-        see TestMulledContainerResolver
-        """
-        cache_directory = os.path.join(self._app.config.container_image_cache_path, "singularity", "mulled")
-        logging.debug(f"TestMulledSingularityContainerResolver.clear_cache {cache_directory}")
-        for filename in os.listdir(cache_directory):
-            file_path = os.path.join(cache_directory, filename)
-            os.unlink(file_path)
-
-    def _assert_container_in_cache(
-        self, cached: bool, container_name: str, namespace: Optional[str] = None, hash_func: Literal["v1", "v2"] = "v2"
-    ) -> None:
-        """
-        see TestMulledContainerResolver
-        """
-        cache_directory = os.path.join(self._app.config.container_image_cache_path, "singularity", "mulled")
-        imageid_list = os.listdir(path=cache_directory)
-        assert cached == (container_name in imageid_list)
 
 
 def disable_dependency_resolution(config: Dict[str, Any]) -> None:
@@ -542,506 +391,3 @@ class TestSingularityJobsIntegration(TestDockerizedJobsIntegration):
     def _assert_container_description_identifier(self, identifier, expected_hash):
         assert os.path.exists(identifier)
         assert identifier.endswith(f"singularity/mulled/{expected_hash}")
-
-
-class ResolverTestProtocol(Protocol):
-    """
-    Helper class defining methods and properties needed to
-    use ContainerResolverTestCases
-    """
-
-    @property
-    def dataset_populator(self) -> DatasetPopulator:
-        ...
-
-    @property
-    def tool_id(self) -> str:
-        ...
-
-    @property
-    def assumptions(self) -> Dict[str, Any]:
-        ...
-
-    @property
-    def container_type(self) -> Literal["docker", "singularity"]:
-        ...
-
-    def _assert_container_in_cache(
-        self, cached: bool, container_name: str, namespace: Optional[str] = None, hash_func: Literal["v1", "v2"] = "v2"
-    ) -> None:
-        ...
-
-    def _assert_status_code_is(self, response: "Response", expected_status_code: int) -> None:
-        ...
-
-    def _get(self, route, data=None, headers=None, admin=False) -> "Response":
-        ...
-
-    def _post(self, route, data=None, files=None, headers=None, admin=False, json: bool = False) -> "Response":
-        ...
-
-
-class ContainerResolverTestCases:
-    """
-    Test cases for (the?) 3 possibilities where Galaxy calls the (container) resolve function
-
-    1. when preparing a job (test_tool_run)
-       - check tool output
-       - check if container has been cached
-    2. when listing container dependencies in the admin UI (test_api_container_resolvers_toolbox)
-       - test consist of 2 calls to the route in order to check if a 2nd round picks
-         up a potentially cached container
-       - after each call check container_type, used resolver and if container has been cached
-    3. when "building" a container in the admin UI (test_api_container_resolvers_toolbox_install)
-       - test consist of 2 calls to the route in order to check if a 2nd round picks
-         up a potentially cached container
-       - after each call check container_type, used resolver and if container has been cached
-
-    Note, for listing and building the test cases set the container type.
-    This is probably not what is done when listing containers in the admin UI,
-    but a container type is only set via the destination a job is executed on.
-    That is, in reality for listing and building both container types are allowed,
-    which might lead to unwanted outcomes if the container resolver config
-    contains resolvers for both container types (like the default does).
-    """
-
-    def test_tool_run(self: ResolverTestProtocol, history_id: str) -> None:
-        """
-        test running a tool
-
-        - runs the tool with tool_id (allowed to fail: assumptions["run"]["expect_failure"])
-        - checks outputs (assumptions["run"]["output"])
-        - check is the container has been cached (assumptions["run"]["cache..."])
-        """
-        try:
-            self.dataset_populator.run_tool(self.tool_id, {}, history_id)
-            self.dataset_populator.wait_for_history(history_id, assert_ok=True, timeout=240)
-        except AssertionError:
-            if not self.assumptions["run"].get("expect_failure", False):
-                raise
-        else:
-            if self.assumptions["run"].get("expect_failure", False):
-                raise AssertionError("test_tool_run is expected to fail")
-        if not self.assumptions["run"].get("expect_failure", False):
-            output = self.dataset_populator.get_history_dataset_content(history_id, timeout=EXTENDED_TIMEOUT)
-            for o in self.assumptions["run"]["output"]:
-                assert o in output
-        self._assert_container_in_cache(
-            self.assumptions["run"]["cached"],
-            container_name=self.assumptions["run"]["cache_name"],
-            namespace=self.assumptions["run"]["cache_namespace"],
-        )
-
-    def _check_status(self: ResolverTestProtocol, status: Dict[str, Any], assumptions:Dict[str, Any]):
-        if "unresolved" in assumptions:
-            assert status["model_class"] == "NullDependency"
-            assert status["dependency_type"] is None
-            assert 'container_resolver' not in status
-            assert 'container_description' not in status
-        else:
-            assert status["model_class"] == "ContainerDependency"
-            assert status["dependency_type"] == self.container_type
-            assert status["container_resolver"]["resolver_type"] == assumptions["resolver_type"]
-            assert re.match(assumptions["identifier"], status["container_description"]["identifier"])
-            self._assert_container_in_cache(
-                assumptions["cached"],
-                container_name=assumptions["cache_name"],
-                namespace=assumptions["cache_namespace"],
-            )
-
-    def test_api_container_resolvers_toolbox(self: ResolverTestProtocol):
-        """
-        test container resolvers via GET container_resolvers/toolbox
-
-        which is what happens when listing containers in the admin UI
-        both calls should resolve with mulled and container is not cached
-
-        test checks assumptions on
-        - resolver type
-        - identifier (a regexp). note for docker the same when cached/uncached
-        - caching (cached, cache_name, cache_namespace)
-        """
-        create_response = self._get(
-            "container_resolvers/toolbox",
-            data={
-                "tool_ids": [self.tool_id],
-                "container_type": self.container_type,
-            },
-            admin=True,
-        )
-        response = create_response.json()
-        assert len(response) == 1
-        status = response[0]["status"]
-        self._check_status(status, self.assumptions["list"][0])
-
-        create_response = self._get(
-            "container_resolvers/toolbox",
-            data={
-                "tool_ids": [self.tool_id],
-                "container_type": self.container_type,
-            },
-            admin=True,
-        )
-        response = create_response.json()
-        assert len(response) == 1
-        status = response[0]["status"]
-        self._check_status(status, self.assumptions["list"][1])
-
-    def test_api_container_resolvers_toolbox_install(self: ResolverTestProtocol):
-        """
-        test container resolvers via POST container_resolvers/toolbox/install
-
-        which is what happens when building a container in the admin UI
-
-        1st call resolves with mulled and container is cached (but the cached
-            URI is returned .. but for docker this makes no difference)
-        2nd call should resolve with cached_mulled (container is still cached)
-        """
-        create_response = self._post(
-            "container_resolvers/toolbox/install",
-            data={
-                "tool_ids": json.dumps([self.tool_id]),
-                "container_type": self.container_type,
-                "include_containers": True,
-            },
-            admin=True,
-        )
-        self._assert_status_code_is(create_response, 200)
-        response = create_response.json()
-        assert len(response) == 1
-        status = response[0]["status"]
-        self._check_status(status, self.assumptions["build"][0])
-
-        create_response = self._post(
-            "container_resolvers/toolbox/install",
-            data={
-                "tool_ids": json.dumps([self.tool_id]),
-                "container_type": self.container_type,
-                "include_containers": True,
-            },
-            admin=True,
-        )
-        self._assert_status_code_is(create_response, 200)
-        response = create_response.json()
-        assert len(response) == 1
-        status = response[0]["status"]
-        self._check_status(status, self.assumptions["build"][1])
-
-
-class TestMulledContainerResolver(ContainerizedIntegrationTestCase, ContainerResolverTestCases):
-    """
-    test cached_mulled + mulled container resolvers in default config
-
-    besides the properties of the returned container description the main points are
-    - running a tool creates an entry in the container cache (the test **can't** show if this is due to
-      1. the `docker inspect .. && docker pull ..` statement in the job script or
-      2. the resolve function
-      it should be both .. see also the corresponding test in TestMulledSingularityContainerResolver where
-      it is definitely the resolve function -- since for singularity there is no additional commands in the
-      job script)
-    - listing containers does not create a cache entry (cached=False and in both calls to resolve mulled is successful)
-    - building the container creates a cache entry (cached=True, 1st call resolves with mulled and 2nd with cached_mulled)
-    """
-
-    tool_id = "mulled_example_multi_1"
-    mulled_hash = "mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0"
-    container_resolvers_config: List[Dict[str, Any]] = [
-        {
-            "type": "cached_mulled",
-        },
-        {"type": "mulled"},
-    ]
-    assumptions: Dict[str, Any] = {
-        "run": {
-            "output": [
-                "bedtools v2.26.0",
-                "samtools: error while loading shared libraries: libcrypto.so.1.0.0",
-            ],
-            "cached": True,
-            "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-            "cache_namespace": "biocontainers",
-        },
-        "list": [
-            {
-                "resolver_type": "mulled",
-                "identifier": "quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "mulled",
-                "identifier": "quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-                "cache_namespace": "biocontainers",
-            },
-        ],
-        "build": [
-            {
-                "resolver_type": "mulled",
-                "identifier": "quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "cached_mulled",
-                "identifier": "quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-                "cache_namespace": "biocontainers",
-            },
-        ],
-    }
-
-    @classmethod
-    def handle_galaxy_config_kwds(cls, config) -> None:
-        super().handle_galaxy_config_kwds(config)
-        config["container_resolvers"] = cls.container_resolvers_config
-
-
-class TestMulledSingularityContainerResolver(SingularityIntegrationTestCase, ContainerResolverTestCases):
-    """
-    test cached_mulled_singularity + mulled_singularity container resolvers in default config
-
-    assumptions:
-    1. tool run
-       - container should still be cached during job preparation (even if the cached image
-         won't be used for the 1st run .. see assumption for building .. would change with auto_install=False)
-    2. listing container
-       - container is not cached
-       - URI is resolved via mulled_singularity (note the `docker://` prefix)
-    3. building container
-       - container is cached in 1st round (via mulled_singularity), but **despite caching the URI is returned**
-       - 2nd round resolves cached image, uses the cached container
-    """
-
-    tool_id = "mulled_example_multi_1"
-    mulled_hash = "mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0"
-    container_resolvers_config: List[Dict[str, Any]] = [
-        {
-            "type": "cached_mulled_singularity",
-        },
-        {
-            "type": "mulled_singularity",
-        },
-    ]
-
-    assumptions = {
-        "run": {
-            "expect_failure": False,
-            "output": [
-                "bedtools v2.26.0",
-                "samtools: error while loading shared libraries: libcrypto.so.1.0.0",
-            ],
-            "cached": True,
-            "cache_name": mulled_hash,
-            "cache_namespace": "biocontainers",
-        },
-        "list": [
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "docker://quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "docker://quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-        ],
-        "build": [
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "docker://quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "cached_mulled_singularity",
-                "identifier": "/tmp/.*/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-        ],
-    }
-
-    @classmethod
-    def handle_galaxy_config_kwds(cls, config) -> None:
-        super().handle_galaxy_config_kwds(config)
-        config["container_resolvers"] = cls.container_resolvers_config
-
-
-class TestMulledContainerResolverNoAutoInstall(TestMulledContainerResolver):
-    """
-    Use the mulled (docker) container resolver with auto_install: False
-
-    No difference (since the cached name is identical to the URI)
-    """
-    pass
-
-
-class TestMulledSingularityContainerResolverNoAutoInstall(TestMulledSingularityContainerResolver):
-    """
-    Use the mulled singularity container resolver with auto_install: False
-
-    The only difference is that the first call to resolve also returns the path
-    to the cached image (see assumptions["build"]["identifier"]). This is also used
-    in the run, but I have no idea how to test this (in the generated job script
-    the path is used instead of the URI)
-    """
-    container_resolvers_config: List[Dict[str, Any]] = [
-        {
-            "type": "cached_mulled_singularity",
-        },
-        {
-            "type": "mulled_singularity",
-            "auto_install": False,
-        },
-    ]
-
-    mulled_hash = "mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0"
-    assumptions = {
-        "run": {
-            "expect_failure": False,
-            "output": [
-                "bedtools v2.26.0",
-                "samtools: error while loading shared libraries: libcrypto.so.1.0.0",
-            ],
-            "cached": True,
-            "cache_name": mulled_hash,
-            "cache_namespace": "biocontainers",
-        },
-        "list": [
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "docker://quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "docker://quay.io/biocontainers/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": False,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-        ],
-        "build": [
-            {
-                "resolver_type": "mulled_singularity",
-                "identifier": "/tmp/.*/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-            {
-                "resolver_type": "cached_mulled_singularity",
-                "identifier": "/tmp/.*/mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0",
-                "cached": True,
-                "cache_name": mulled_hash,
-                "cache_namespace": "biocontainers",
-            },
-        ],
-    }
-
-
-class TestCondaFallBack(ContainerizedIntegrationTestCase, ContainerResolverTestCases):
-    """
-    test that Galaxy falls back to default dependency resolvers (i.e. conda) if no
-    container can be resolved
-
-    here we force container resolution to fail because only singularity resolvers
-    are configured on a docker destination.
-
-    - assumptions that need to be met is that tool is executed successfully (via conda)
-    - and listing and building containers does not work
-    """
-
-    allow_conda_fallback: bool = True
-    tool_id = "mulled_example_multi_1"
-    mulled_hash = "mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0"
-    container_resolvers_config: List[Dict[str, Any]] = [
-        {"type": "cached_mulled_singularity"},
-        {"type": "mulled_singularity"},
-    ]
-
-    assumptions: Dict[str, Any] = {
-        "run": {
-            "output": [
-                "bedtools v2.26.0",
-                # conda env does not suffer from broken library -> so different output
-                "samtools (Tools for alignments in the SAM format)",
-                "Version: 1.3.1",
-            ],
-            "cached": False,
-            "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-            "cache_namespace": "biocontainers",
-        },
-        "list": [
-            {"unresolved": True},
-            {"unresolved": True},
-        ],
-        "build": [
-            {"unresolved": True},
-            {"unresolved": True},
-        ],
-    }
-
-    @classmethod
-    def handle_galaxy_config_kwds(cls, config) -> None:
-        super().handle_galaxy_config_kwds(config)
-        config["container_resolvers"] = cls.container_resolvers_config
-
-
-class TestCondaFallBackAndRequireContainer(ContainerizedIntegrationTestCase, ContainerResolverTestCases):
-    """
-    test that we can disable fallback to the default resolvers (conda)
-    by setting the destination property require_container
-
-    same as TestCondaFallBack but tool needs to fail
-    """
-
-    allow_conda_fallback: bool = True
-    tool_id = "mulled_example_multi_1"
-    mulled_hash = "mulled-v2-8186960447c5cb2faa697666dc1e6d919ad23f3e:a6419f25efff953fc505dbd5ee734856180bb619-0"
-    container_resolvers_config: List[Dict[str, Any]] = [
-        {"type": "cached_mulled_singularity"},
-        {"type": "mulled_singularity"},
-    ]
-
-    assumptions: Dict[str, Any] = {
-        "run": {
-            "output": [
-                "bedtools v2.26.0",
-                # conda env does not suffer from broken library -> so different output
-                "samtools (Tools for alignments in the SAM format)",
-                "Version: 1.3.1",
-            ],
-            "cached": False,
-            "cache_name": f"quay.io/biocontainers/{mulled_hash}",
-            "cache_namespace": "biocontainers",
-        },
-        "list": [
-            {"unresolved": True},
-            {"unresolved": True},
-        ],
-        "build": [
-            {"unresolved": True},
-            {"unresolved": True},
-        ],
-    }
-
-    @classmethod
-    def handle_galaxy_config_kwds(cls, config) -> None:
-        super().handle_galaxy_config_kwds(config)
-        config["job_config"]["execution"]["environments"]["local_docker"]["require_container"] = True
-        config["container_resolvers"] = cls.container_resolvers_config
